+++ tracdap-libs/tracdap-lib-data/src/main/java/org/finos/tracdap/common/async/flow/SourcePublisher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Accenture Global Solutions Limited
+ * Copyright 2023 Accenture Global Solutions Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.concurrent.flow;
+package org.finos.tracdap.common.async.flow;
 
 import org.finos.tracdap.common.exception.ETracInternal;
