+++ tracdap-libs/tracdap-lib-data/src/test/java/org/finos/tracdap/common/storage/StorageReadWriteTestSuite.java
@@ -20,7 +20,7 @@ import org.finos.tracdap.common.data.IDataContext;
 import org.finos.tracdap.common.data.util.Bytes;
 import org.finos.tracdap.common.exception.EStorageRequest;
 import org.finos.tracdap.common.exception.EValidationGap;
-import org.finos.tracdap.common.concurrent.Flows;
+import org.finos.tracdap.common.async.Flows;
 
 import org.apache.arrow.memory.ArrowBuf;
