+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/DataApiTestHelpers.java
@@ -21,7 +21,7 @@ import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.exception.EUnexpected;
 import org.finos.tracdap.common.metadata.MetadataCodec;
 import org.finos.tracdap.common.metadata.MetadataUtil;
-import org.finos.tracdap.common.concurrent.Flows;
+import org.finos.tracdap.common.async.Flows;
 import org.finos.tracdap.common.data.util.ByteSeekableChannel;
 import org.finos.tracdap.test.grpc.GrpcTestStreams;
 import org.finos.tracdap.metadata.BasicType;
