+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/DataOperationsTest.java
@@ -18,7 +18,7 @@ package org.finos.tracdap.svc.data.api;
 
 import org.finos.tracdap.api.*;
 import org.finos.tracdap.common.data.ExecutionContext;
-import org.finos.tracdap.common.concurrent.Futures;
+import org.finos.tracdap.common.async.Futures;
 import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.metadata.MetadataCodec;
