+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/FileOperationsTest.java
@@ -22,8 +22,8 @@ import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.metadata.MetadataCodec;
 import org.finos.tracdap.common.metadata.MetadataUtil;
-import org.finos.tracdap.common.concurrent.Flows;
-import org.finos.tracdap.common.concurrent.Futures;
+import org.finos.tracdap.common.async.Flows;
+import org.finos.tracdap.common.async.Futures;
 import org.finos.tracdap.common.plugin.PluginManager;
 import org.finos.tracdap.metadata.*;
